silence warnings

+++ sources/Application/Views/BaseClasses/View.h
@@ -202,8 +202,10 @@ protected:
     // Map dB to bar levels  -60dB to 0dB range mapped to 0-159 bars
     // Optimized 159/60 ≈ 2.65 = (2.65 * 256) / 256 = 678 / 256
     // Using fixed-point: multiply by 678, then right-shift by 8 (divide by 256)
-    *left = std::max(0, std::min(VU_METER_MAX, ((leftDb + 60) * 678) >> 8));
-    *right = std::max(0, std::min(VU_METER_MAX, ((rightDb + 60) * 678) >> 8));
+    *left = std::max(
+        0, std::min(VU_METER_MAX, (int32_t)(((leftDb + 60) * 678) >> 8)));
+    *right = std::max(
+        0, std::min(VU_METER_MAX, (int32_t)(((rightDb + 60) * 678) >> 8)));
   }
 
 public: // temp hack for modl windo constructors
